// vision-line-webapp/src/api/models/Function.ts
// tslint:disable
/**
 * Vision Line API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import {
    FunctionInput,
} from './';

/**
 * @export
 * @interface Function
 */
export interface Function {
    /**
     * @type {string}
     * @memberof Function
     */
    name?: string;
    /**
     * @type {string}
     * @memberof Function
     */
    display?: string;
    /**
     * @type {Array<FunctionInput>}
     * @memberof Function
     */
    inputs?: Array<FunctionInput>;
}