// vision-line-webapp/src/api/models/Node.ts
// tslint:disable
/**
 * Vision Line API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * @export
 * @interface Node
 */
export interface Node {
    /**
     * ID of Node
     * @type {number}
     * @memberof Node
     */
    id?: number;
    /**
     * Name of the implemented function
     * @type {string}
     * @memberof Node
     */
    name?: string;
    /**
     * Position of node in line
     * @type {number}
     * @memberof Node
     */
    position?: number;
    /**
     * Input key-value pairs as object
     * @type {object}
     * @memberof Node
     */
    inputs?: object;
}